Combined usertracking and heartbeat sender plugin
